feat: basic plugins
