+++ nvim/.gitignore
@@ -1,1 +1,1 @@
-plugin/packer_compiled.lua
+lazy-lock.json
