--- nvim/after/plugin/colors.lua
@@ -1,14 +1,0 @@
-require('rose-pine').setup({
-  disable_background = true
-})
-
-function ColorMyPencils(color)
-	color = color or "rose-pine"
-	vim.cmd.colorscheme(color)
-
-	vim.api.nvim_set_hl(0, "Normal", { bg = "none" })
-  vim.api.nvim_set_hl(0, "NormalFloat", { bg = "none" })
-  vim.api.nvim_set_hl(0, "SignColumn", { bg = "none" })
-end
-
-ColorMyPencils()
