--- nvim/after/plugin/lualine.lua
@@ -1,5 +1,0 @@
-require('lualine').setup({
-  sections = {
-    lualine_c = { {'filename', path = 1}  }
-  },
-})
