--- nvim/after/plugin/which-key.lua
@@ -1,1 +1,0 @@
-require("which-key").setup {}
