+++ nvim/lua/hackder/plugins/autopairs.lua
@@ -1,0 +1,8 @@
+return {
+  {
+    "windwp/nvim-autopairs",
+    config = function()
+      require("nvim-autopairs").setup({})
+    end
+  }
+}
