+++ nvim/lua/hackder/plugins/colorscheme.lua
@@ -1,0 +1,17 @@
+return {
+  "folke/tokyonight.nvim",
+  {
+    "rose-pine/neovim",
+    as = "rose-pine",
+    config = function()
+      require("rose-pine").setup({
+        disable_background = true
+      })
+
+      vim.cmd("colorscheme rose-pine")
+      vim.api.nvim_set_hl(0, "Normal", { bg = "none" })
+      vim.api.nvim_set_hl(0, "NormalFloat", { bg = "none" })
+      vim.api.nvim_set_hl(0, "SignColumn", { bg = "none" })
+    end
+  }
+}
