+++ nvim/lua/hackder/plugins/copilot.lua
@@ -1,0 +1,13 @@
+return {
+  {
+    "github/copilot.vim",
+    config = function()
+      vim.g.copilot_node_command = "node"
+      vim.g.copilot_no_tab_map = true
+      vim.g.copilot_assume_mapped = true
+
+      vim.keymap.set("i", "<C-j>", 'copilot#Accept("<CR>")',
+        { noremap = true, silent = true, expr = true, replace_keycodes = false, desc = 'Copilot accept' })
+    end
+  }
+}
