+++ nvim/lua/hackder/plugins/format-ts-errors.lua
@@ -1,0 +1,5 @@
+return {
+  {
+    "davidosomething/format-ts-errors.nvim"
+  }
+}
