+++ nvim/lua/hackder/plugins/init.lua
@@ -1,0 +1,5 @@
+return {
+  "nvim-lua/plenary.nvim",
+  "folke/neodev.nvim",
+  { "folke/neoconf.nvim", cmd = "Neoconf" },
+}
