+++ nvim/lua/hackder/plugins/lualine.lua
@@ -1,0 +1,12 @@
+return {
+  {
+    "nvim-lualine/lualine.nvim",
+    config = function()
+      require('lualine').setup({
+        sections = {
+          lualine_c = { {'filename', path = 1}  }
+        },
+      })
+    end
+  }
+}
