+++ nvim/lua/hackder/plugins/treesitter.lua
@@ -1,0 +1,184 @@
+return {
+  {
+    "nvim-treesitter/nvim-treesitter",
+    build = ":TSUpdate",
+    event = { "VeryLazy" },
+    init = function(plugin)
+      require("lazy.core.loader").add_to_rtp(plugin)
+      require("nvim-treesitter.query_predicates")
+    end,
+    config = function()
+      require('nvim-treesitter.configs').setup({
+        -- A list of parser names, or "all"
+        ensure_installed = { "javascript", "typescript", "python", "cpp", "lua", "rust" },
+
+        -- Install parsers synchronously (only applied to `ensure_installed`)
+        sync_install = false,
+
+        -- Automatically install missing parsers when entering buffer
+        -- Recommendation: set to false if you don't have `tree-sitter` CLI installed locally
+        auto_install = true,
+
+        ---- If you need to change the installation directory of the parsers (see -> Advanced Setup)
+        -- parser_install_dir = "/some/path/to/store/parsers", -- Remember to run vim.opt.runtimepath:append("/some/path/to/store/parsers")!
+
+        highlight = {
+          -- `false` will disable the whole extension
+          enable = true,
+
+          -- Setting this to true will run `:h syntax` and tree-sitter at the same time. Set this to `true` if you depend on 'syntax' being enabled (like for indentation).
+          -- Using this option may slow down your editor, and you may see some duplicate highlights.
+          -- Instead of true it can also be a list of languages
+          additional_vim_regex_highlighting = false,
+        },
+        textobjects = {
+          select = {
+            enable = true,
+            lookahead = true, -- Automatically jump forward to textobj, similar to targets.vim
+            keymaps = {
+              -- You can use the capture groups defined in textobjects.scm
+              ['aa'] = '@parameter.outer',
+              ['ia'] = '@parameter.inner',
+              ['af'] = '@function.outer',
+              ['if'] = '@function.inner',
+              ['ac'] = '@class.outer',
+              ['ic'] = '@class.inner',
+              ['ii'] = '@conditional.inner',
+              ['ai'] = '@conditional.outer',
+              ['il'] = '@loop.inner',
+              ['al'] = '@loop.outer',
+              ['at'] = '@comment.outer',
+            },
+          },
+          move = {
+            enable = true,
+            set_jumps = true, -- whether to set jumps in the jumplist
+            goto_next_start = {
+              [']]'] = '@function.outer',
+              [']c'] = '@class.outer',
+            },
+            goto_next_end = {
+              [']['] = '@function.outer',
+              [']C'] = '@class.outer',
+            },
+            goto_previous_start = {
+              ['[['] = '@function.outer',
+              ['[c'] = '@class.outer',
+            },
+            goto_previous_end = {
+              ['[['] = '@function.outer',
+              ['[C'] = '@class.outer',
+            },
+            -- goto_next = {
+            --   [']i'] = "@conditional.inner",
+            -- },
+            -- goto_previous = {
+            --   ['[i'] = "@conditional.inner",
+            -- }
+          },
+        },
+      })
+    end
+  },
+  {
+    "nvim-treesitter/nvim-treesitter-context",
+    config = function()
+      require'treesitter-context'.setup{
+        enable = true, -- Enable this plugin (Can be enabled/disabled later via commands)
+        max_lines = 0, -- How many lines the window should span. Values <= 0 mean no limit.
+        trim_scope = 'outer', -- Which context lines to discard if `max_lines` is exceeded. Choices: 'inner', 'outer'
+        min_window_height = 0, -- Minimum editor window height to enable context. Values <= 0 mean no limit.
+        patterns = { -- Match patterns for TS nodes. These get wrapped to match at word boundaries.
+            -- For all filetypes
+            -- Note that setting an entry here replaces all other patterns for this entry.
+            -- By setting the 'default' entry below, you can control which nodes you want to
+            -- appear in the context window.
+            default = {
+                'class',
+                'function',
+                'method',
+                'for',
+                'while',
+                'if',
+                'switch',
+                'case',
+                'interface',
+                'struct',
+                'enum',
+            },
+            -- Patterns for specific filetypes
+            -- If a pattern is missing, *open a PR* so everyone can benefit.
+            tex = {
+                'chapter',
+                'section',
+                'subsection',
+                'subsubsection',
+            },
+            haskell = {
+                'adt'
+            },
+            rust = {
+                'impl_item',
+
+            },
+            terraform = {
+                'block',
+                'object_elem',
+                'attribute',
+            },
+            scala = {
+                'object_definition',
+            },
+            vhdl = {
+                'process_statement',
+                'architecture_body',
+                'entity_declaration',
+            },
+            markdown = {
+                'section',
+            },
+            elixir = {
+                'anonymous_function',
+                'arguments',
+                'block',
+                'do_block',
+                'list',
+                'map',
+                'tuple',
+                'quoted_content',
+            },
+            json = {
+                'pair',
+            },
+            typescript = {
+                'export_statement',
+            },
+            yaml = {
+                'block_mapping_pair',
+            },
+        },
+        exact_patterns = {
+            -- Example for a specific filetype with Lua patterns
+            -- Treat patterns.rust as a Lua pattern (i.e "^impl_item$" will
+            -- exactly match "impl_item" only)
+            -- rust = true,
+        },
+
+        -- [!] The options below are exposed but shouldn't require your attention,
+        --     you can safely ignore them.
+
+        zindex = 20, -- The Z-index of the context window
+        mode = 'cursor',  -- Line used to calculate context. Choices: 'cursor', 'topline'
+        -- Separator between context and content. Should be a single character string, like '-'.
+        -- When separator is set, the context will only show up when there are at least 2 lines above cursorline.
+        separator = nil,
+    }
+    end
+  },
+  {
+    "nvim-treesitter/nvim-treesitter-textobjects",
+    dependencies = {
+      "nvim-treesitter/nvim-treesitter"
+    }
+  }
+}
