+++ nvim/lua/hackder/plugins/which-key.lua
@@ -1,0 +1,3 @@
+  return {
+    "folke/which-key.nvim",
+  }
